backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
python-dotenv>=1.0,<2.0
Pillow>=10.4,<11.0
django-storages[boto3]>=1.14,<1.15
psycopg2-binary>=2.9,<3.0
mysqlclient>=2.2,<3.0
celery[redis]>=5.4,<5.5
redis>=5.0.0,<5.1.0
kombu>=5.4,<5.5
flower>=2.0,<3.0
opencv-python>=4.8.0,<5.0.0
scikit-image>=0.21.0,<0.23.0
scipy>=1.10.0,<2.0.0
numpy>=1.24.0,<2.0.0





